change organizationActivity to a dir and move related components to be inside
